Refactor integration tests and add more tests
