+++ Gemfile
@@ -12,8 +12,8 @@ gem "coveralls", "~> 0.7"
 gem "yard", "~> 0.9"
 gem "yard-doctest", "~> 0.1.8"
 gem "gems", "~> 0.8"
-gem "actionpack", ">= 4.0"
-gem "railties", ">= 4.0"
+gem "actionpack", "~> 4.0"
+gem "railties", "~> 4.0"
 gem "rack", ">= 0.1"
 gem "sinatra", "~> 1.4"
