+++ google-cloud-error_reporting/Rakefile
@@ -58,6 +58,7 @@ namespace :integration do
     ENV["TEST_GOOGLE_CLOUD_PROJECT_URI"] = args[:project_uri]
 
     $LOAD_PATH.unshift "lib", "integration"
+    Dir.glob("integration/*_test.rb").each { |file| require_relative file }
     Dir.glob("integration/gae/**/*_test.rb").each { |file| require_relative file }
   end
 
@@ -69,6 +70,7 @@ namespace :integration do
     ENV["TEST_GKE_POD_NAME"] = args[:pod_name]
 
     $LOAD_PATH.unshift "lib", "integration"
+    Dir.glob("integration/*_test.rb").each { |file| require_relative file }
     Dir.glob("integration/gke/**/*_test.rb").each { |file| require_relative file }
   end
 end
