+++ google-cloud-error_reporting/google-cloud-error_reporting.gemspec
@@ -30,7 +30,7 @@ Gem::Specification.new do |gem|
   gem.add_development_dependency "minitest-autotest", "~> 1.0"
   gem.add_development_dependency "minitest-focus", "~> 1.1"
   gem.add_development_dependency "minitest-rg", "~> 5.2"
-  gem.add_development_dependency "actionpack", ">= 4.0"
-  gem.add_development_dependency "railties", ">= 4.0"
+  gem.add_development_dependency "actionpack", "~> 4.0"
+  gem.add_development_dependency "railties", "~> 4.0"
   gem.add_development_dependency "rack", ">= 0.1"
 end
