+++ google-cloud-logging/integration/gke/logging_test.rb
@@ -13,38 +13,15 @@
 # limitations under the License.
 
 
-require "minitest/autorun"
-require "minitest/rg"
-require "minitest/focus"
-require "net/http"
-require "open3"
-require "json"
+require "logging_helper"
 require "google/cloud/logging"
-require_relative "../../../integration/helper"
 
 describe Google::Cloud::Logging do
-  let(:gke_pod_name) { ENV["TEST_GKE_POD_NAME"] }
+  it "Uses monitored resource with 'container' type" do
+    response = JSON.parse send_request("test_logger")
 
-  it "submits log event to Stackdriver Logging service" do
-    token = Time.now.to_i
-    logging_uri = URI("http://localhost:8080/test_logging")
-    logging_uri.query="token=#{token}"
-
-    stdout = Open3.capture3("gcloud config list project").first
-    project_id = stdout.scan(/project = (.*)/).first.first
-
-    `kubectl exec #{gke_pod_name} -- curl #{logging_uri.to_s}`
-
-    logs = []
-    keep_trying_till_true do
-      stdout = Open3.capture3(
-        "gcloud beta logging read \"logName=projects/#{project_id}/logs/google-cloud-ruby_integration_test " \
-          "AND textPayload:#{token}\" --limit 1 --format json"
-      ).first
-      logs = JSON.parse(stdout)
-      logs.length == 1
-    end
-
-    logs.length.must_equal 1
+    response["monitored_resource"]["type"].must_equal "container"
+    response["monitored_resource"]["labels"]["cluster_name"].wont_be_nil
+    response["monitored_resource"]["labels"]["namespace_id"].wont_be_nil
   end
 end
