+++ google-cloud/google-cloud.gemspec
@@ -41,7 +41,4 @@ Gem::Specification.new do |gem|
   gem.add_development_dependency "simplecov", "~> 0.9"
   gem.add_development_dependency "yard", "~> 0.9"
   gem.add_development_dependency "yard-doctest", "~> 0.1.8"
-  gem.add_development_dependency "actionpack", ">= 4.0"
-  gem.add_development_dependency "railties", ">= 4.0"
-  gem.add_development_dependency "rack", ">= 0.1"
 end
