--- integration/rails4_app/app/views/test/test.html.erb
@@ -1,5 +1,0 @@
-<% (@outputs || []).each do |output| %>
-	<div>
-		<%= output %>
-	</div>
-<% end %>
