+++ integration/rails4_app/config/routes.rb
@@ -3,4 +3,5 @@ Rails.application.routes.draw do
 
   get '/test_error_reporting', to: 'test#test_error_reporting'
   get '/test_logging', to: 'test#test_logging'
+  get '/test_logger', to: 'test#test_logger'
 end
